requirements.txt
flask
python-dotenv
markdown
pygments
openai-whisper
openai
